requirements.txt
fastapi==0.109.2
uvicorn==0.27.1
celery==5.3.6
redis==5.0.1
gspread==5.12.4
google-auth==2.27.0
pydantic==2.6.1
pydantic-settings==2.1.0
python-json-logger==2.0.7
prometheus-client==0.19.0
supabase==2.3.1
httpx>=0.24,<0.26
pytz==2024.1
opentelemetry-api==1.22.0
opentelemetry-sdk==1.22.0
opentelemetry-instrumentation-fastapi==0.44b0
opentelemetry-instrumentation-redis==0.44b0
opentelemetry-instrumentation-requests==0.44b0
opentelemetry-exporter-jaeger==1.22.0
opentelemetry-exporter-otlp==1.22.0
psutil==5.9.8
